+++ R/read.R
@@ -65,6 +65,7 @@
 #' @param keep_attributes If `TRUE` additional attributes are returned.
 #'   (These are used internally to define a cell type.)
 #' @param check_names If `TRUE` then the names of the variables in the data frame are checked to ensure that they are syntactically valid variable names.
+#' @param show_hyperlinks If `TRUE` instead of the displayed text, hyperlink targets are shown.
 #' @param ... additional arguments
 #'
 #' @examples
@@ -156,6 +157,7 @@ wb_to_df <- function(
     named_region,
     keep_attributes   = FALSE,
     check_names       = FALSE,
+    show_hyperlinks = FALSE,
     ...
 ) {
 
@@ -177,8 +179,12 @@ wb_to_df <- function(
     if (missing(sheet))
       sheet <- substitute()
 
+    data_only <- TRUE
+    # TODO hyperlinks are deeper embedded into the wb_load code
+    if (show_hyperlinks) data_only <- FALSE
+
     # possible false positive on current lintr runs
-    wb <- wb_load(file, sheet = sheet, data_only = TRUE) # nolint
+    wb <- wb_load(file, sheet = sheet, data_only = data_only) # nolint
   }
 
   if (!missing(named_region)) {
@@ -470,6 +476,21 @@ wb_to_df <- function(
 
   }
 
+  if (show_hyperlinks) {
+
+    if (length(wb$worksheets[[sheet]]$hyperlinks)) {
+      hyprlnks <- as.data.frame(
+        do.call("rbind",
+                lapply(wb$worksheets[[sheet]]$hyperlinks, function(hl) {
+                  c(hl$ref, ifelse(is.null(hl$target), hl$location, hl$target))
+                })
+        )
+      )
+      cc$val[match(hyprlnks$V1, cc$r)] <- hyprlnks$V2
+    }
+
+  }
+
   # convert "na_string" to missing
   if (has_na_string) cc$typ[cc$typ == "na_string"] <- NA
   if (has_na_number) cc$typ[cc$typ == "na_number"] <- NA
@@ -684,6 +705,7 @@ read_xlsx <- function(
   na.numbers        = NA,
   fill_merged_cells = FALSE,
   check_names       = FALSE,
+  show_hyperlinks   = FALSE,
   ...
 ) {
 
@@ -712,7 +734,8 @@ read_xlsx <- function(
     na.numbers        = na.numbers,
     fill_merged_cells = fill_merged_cells,
     check_names       = check_names,
-    ...
+    show_hyperlinks   = show_hyperlinks,
+    ...               = ...
   )
 }
 
@@ -735,6 +758,7 @@ wb_read <- function(
   na.strings      = "NA",
   na.numbers      = NA,
   check_names     = FALSE,
+  show_hyperlinks = FALSE,
   ...
 ) {
 
@@ -762,7 +786,8 @@ wb_read <- function(
     na.strings      = na.strings,
     na.numbers      = na.numbers,
     check_names     = check_names,
-    ...
+    show_hyperlinks = show_hyperlinks,
+    ...             = ...
   )
 
 }